move useAddress and AddressInput into separate file
